Script for spawning tiles over Grid Map
The tiles are child nodes that contain a flat mesh with a square texture on it. These highlight the walkable map area. Will finagle some way to make it interactible by makingit change colour when hovering with  mouse, and for an additional branch of tiles to show permitter travel distance from a movement position.
There is work to be done, but it is going good once I get the brain juices flowing.
END NOTE