Implemented Event Bus
I now have an event bus that is at the very top (autoloaded) that I can emit signals via so that they can communicate globally. I am using this to emit signals from the tile grid that characters are attached to so that I can detect which specific tile was clicked and retrieve their x and z position, and therefore which tile in the array of tiles it is.

This was so bloody difficult to implement because I had to go on faith that it would work as not a single damn example of this indicated it would work with objects that are instantiatied during run-time. This is hard. I tried.